backend/requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0.0
asyncpg>=0.27.0
aiomqtt>=1.0.0 